# GridExpand/docs/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PipelineBody": {
        "additionalProperties": false,
        "description": "A synthetic pipeline run for the grids of one region.",
        "properties": {
          "ags": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "AGS; resolved from the PLZ when omitted",
            "title": "Ags"
          },
          "candidate_indexes": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "consecutive candidate numbers of the AGS",
            "title": "Candidate Indexes"
          },
          "grid_result_id": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "exactly this pylovo grid (ignores min_buildings)",
            "title": "Grid Result Id"
          },
          "min_buildings": {
            "default": 5,
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Min Buildings",
            "type": "integer"
          },
          "model_cases": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Model Cases",
            "type": "array"
          },
          "plz": {
            "anyOf": [
              {
                "maximum": 99999.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "only the grids of this PLZ",
            "title": "Plz"
          },
          "powerflow_output": {
            "default": "summary",
            "pattern": "^(summary|both)$",
            "title": "Powerflow Output",
            "type": "string"
          },
          "pylovo_version_id": {
            "pattern": "^[\\w.-]{1,10}$",
            "title": "Pylovo Version Id",
            "type": "string"
          },
          "scenario": {
            "description": "file name from GET /api/scenarios",
            "maxLength": 200,
            "minLength": 1,
            "title": "Scenario",
            "type": "string"
          },
          "timeframe_mode": {
            "default": "max_base_electricity_demand_week",
            "title": "Timeframe Mode",
            "type": "string"
          },
          "workers": {
            "default": 1,
            "maximum": 16.0,
            "minimum": 1.0,
            "title": "Workers",
            "type": "integer"
          }
        },
        "required": [
          "pylovo_version_id",
          "scenario"
        ],
        "title": "PipelineBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Jobs, results and scenario files of GridExpand (load allocation, urbs optimisation, power flow, grid expansion) for the GridPlanner UI. API version 1.",
    "title": "GridExpand API",
    "version": "api-1",
    "x-api-version": 1
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs/pipeline": {
      "post": {
        "description": "Queue ``gridexpand run`` for each selected model case (one after another).",
        "operationId": "start_pipeline_api_jobs_pipeline_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Start Pipeline Api Jobs Pipeline Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Pipeline",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/terminal": {
      "get": {
        "description": "Runs prepared for a terminal (newest first) with the state of their run directory.",
        "operationId": "terminal_runs_api_jobs_terminal_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Terminal Runs Api Jobs Terminal Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Terminal Runs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "description": "Write the run YAML of a request for a run in a terminal (``tmux``) instead of a job.\n\nReturns the YAML, the commands to start and follow it, and a portable copy of the YAML\n(scenario referenced as ``../scenarios/<file>``) for another GridExpand checkout.",
        "operationId": "prepare_terminal_run_api_jobs_terminal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Prepare Terminal Run Api Jobs Terminal Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prepare Terminal Run",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "description": "One job with the per-grid status of every step (and the last ``tail`` log lines).",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tail",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 5000,
              "minimum": 0,
              "title": "Tail",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Job Api Jobs  Job Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "description": "Cancel a queued job, or stop a running one (SIGTERM to its process group).",
        "operationId": "cancel_job_api_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Job Api Jobs  Job Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/events": {
      "get": {
        "description": "Server-Sent Events: ``log`` (new lines), ``status`` (summary changes), final ``end``.",
        "operationId": "job_events_api_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "title": "After",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Events",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/files": {
      "get": {
        "description": "Files in the job's run directories (events, status, summaries, per-grid logs).",
        "operationId": "job_files_api_jobs__job_id__files_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Job Files Api Jobs  Job Id  Files Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Files",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/files/{path}": {
      "get": {
        "description": "One file of the job's run directories (text files only).",
        "operationId": "job_file_api_jobs__job_id__files__path__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job File",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/log": {
      "get": {
        "description": "Log lines after a sequence number (``format=text`` downloads the whole log).",
        "operationId": "job_log_api_jobs__job_id__log_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Log",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}